+++ controllers/movimiento_cuenta_corriente/index.js
@@ -1,60 +1,72 @@
-//importa la libreria express q permite hacer todo bonito y facil
 import express from "express";
-//importa el modelo perfil donde se definio esa entidad
-import { MovCuentaCorriente } from "../../models/index.js";
+import { Cuenta_corriente, MovCuentaCorriente } from "../../models/index.js";
 
-//crea una ruta para la entidad
 export const movimiento_cuenta_corrienteRouter = express.Router();
 
-//crea ruta que crea un perfil
 movimiento_cuenta_corrienteRouter.post("/", async (req, res) => {
-  const { hora, monto } = req.body;
+  const { hora, monto, nroCuenta } = req.body;
   if (hora.length < 0)
-    return res.status(400).send({ message: "la hora no coincide " });
+    return res.status(400).send({ message: "la hora no puede llegar vacia" });
   if (monto.length < 0)
-    return res.status(400).send({ message: "el monto no coincide " });
-  const result = await MovCuentaCorriente.create({ nombre });
-  res.status(201).json({
+    return res.status(400).send({ message: "el monto no puede llegar vacio" });
+  if (!(await Cuenta_corriente.findByPk(nroCuenta)))
+    return res
+      .status(400)
+      .send({ message: "cuenta corriente enviado no fue encontrado" });
+  const movimiento = await MovCuentaCorriente.create({
+    hora,
+    monto,
+    cuentaCorrienteNroCuenta: nroCuenta,
+  });
+  return res.status(201).json({
     message: "movimiento cuenta corriente creado exitosamente",
-    data: { movimientos: result },
+    data: { movimiento },
   });
 });
 
-//crea ruta que obtiene todos los perfiles
 movimiento_cuenta_corrienteRouter.get("/", async (req, res) => {
   const result = await MovCuentaCorriente.findAll();
-  res.status(200).json({
+  return res.status(200).json({
     message: "movimiento cuenta corriente obtenidos exitosamente",
     data: { movimientos: result },
   });
 });
 
-//crea ruta que obtiene un unico perfil
 movimiento_cuenta_corrienteRouter.get("/:id", async (req, res) => {
   const result = await MovCuentaCorriente.findByPk(req.params.id);
+  if (!result)
+    return res.status(400).send({
+      message: " movimiento de cuenta corriente enviado no fue encontrado",
+    });
   res.status(200).json({
     message: "movimiento cuenta corriente obtenido exitosamente",
     data: { movimientos: result },
   });
 });
 
-//crea ruta que borra un perfil
 movimiento_cuenta_corrienteRouter.delete("/:id", async (req, res) => {
+  if (!(await MovCuentaCorriente.findByPk(req.params.id)))
+    return res.status(400).send({
+      message: "movimiento de cuenta corriente enviado no fue encontrado",
+    });
   await MovCuentaCorriente.destroy({
     where: { id: req.params.id },
   });
   res.status(204);
 });
 
-//crea ruta que modifica un perfil
 movimiento_cuenta_corrienteRouter.put("/:id", async (req, res) => {
-  const { hora, monto } = req.body;
+  if (!(await MovCuentaCorriente.findByPk(req.params.id)))
+    return res.status(400).send({
+      message: "movimiento de cuenta corriente enviado no fue encontrado",
+    });
+  const { hora, monto, nroCuenta } = req.body;
   if (hora.length < 0)
     return res.status(400).send({ message: "la hora no existe " });
   if (monto.length < 0)
     return res.status(400).send({ message: "el monto no existe" });
   const result = await MovCuentaCorriente.update(
-    { hora, monto },
+    { hora, monto, cuentaCorrienteNroCuenta: nroCuenta },
     {
       where: { id: req.params.id },
       returning: true,
